+++ apps/src/dance/ai/DanceAiModal.tsx
@@ -2,55 +2,20 @@ import React, {useState, useEffect} from 'react';
 import moduleStyles from './dance-ai-modal.module.scss';
 import AccessibleDialog from '@cdo/apps/templates/AccessibleDialog';
 import Button from '@cdo/apps/templates/Button';
-import HttpClient from '@cdo/apps/util/HttpClient';
-import {CHAT_COMPLETION_URL} from '@cdo/apps/aichat/constants';
 import {useSelector} from 'react-redux';
 import {useAppDispatch} from '@cdo/apps/util/reduxHooks';
-import {setCurrentAiModalField, DanceState} from './danceRedux';
+import {setCurrentAiModalField, DanceState} from '../danceRedux';
 import {StrongText, Heading5} from '@cdo/apps/componentLibrary/typography';
 import classNames from 'classnames';
 import FontAwesome from '@cdo/apps/templates/FontAwesome';
-import {AiModalItem, AiModalReturnedItem} from './types';
+import {AiModalItem, AiModalReturnedItem} from '../types';
 import {BlockSvg} from 'blockly/core';
+import {doAi} from './utils';
 const Typist = require('react-typist').default;
 
 const aiBotBorder = require('@cdo/static/dance/ai-bot-border.png');
 
 const promptString = 'Generate a scene using this mood:';
-
-const doAi = async (input: string) => {
-  const systemPrompt =
-    'You are a helper which can accept a request for a mood or atmosphere, and you then generate JSON like the following format: {backgroundColor: "black", backgroundEffect: "splatter", foregroundEffect: "rain", dancers: {type: "frogs", count: 3, layout: "circle"}}.  The only valid values for backgroundEffect are circles, color_cycle, diamonds, disco_ball, fireworks, swirl, kaleidoscope, lasers, splatter, rainbow, snowflakes, galaxy, sparkles, spiral, disco, stars.  The only valid values for backgroundColor are rave, cool, electronic, iceCream, neon, tropical, vintage, warm.  The only valid values for foregroundEffect are bubbles, confetti, hearts_red, music_notes, pineapples, pizzas, smiling_poop, rain, floating_rainbows, smile_face, spotlight, color_lights, raining_tacos.  Make sure you always generate all three of these values.  The dancers sub-object is optional, but when included, it has three parameters of its own: type must be one of the following: alien, bear, cat, dog, duck, frog, moose, pineapple, robot, shark, sloth, unicorn; count must be between 2 and 40 inclusive; layout must be one of the following: border, inner, diamond, circle, grid, top, row, bottom, left, column, right, x, plus, spiral, random.   Also, add a field called "explanation" to the result JSON, which contains a simple explanation of why you chose the values that you did, at the reading level of a 5th-grade school student, in one sentence of less than forty words.';
-
-  const messages = [
-    {
-      role: 'system',
-      content: systemPrompt,
-    },
-    {
-      role: 'user',
-      content: input,
-    },
-  ];
-
-  console.log('do AI:', input);
-
-  const response = await HttpClient.post(
-    CHAT_COMPLETION_URL,
-    JSON.stringify({messages}),
-    true,
-    {
-      'Content-Type': 'application/json; charset=UTF-8',
-    }
-  );
-
-  if (response.status === 200) {
-    const res = await response.json();
-    return res.content;
-  } else {
-    return null;
-  }
-};
 
 export enum Mode {
   SELECT_INPUTS = 'selectInputs',
