+++ apps/src/dance/blocks.js
@@ -1,5 +1,5 @@
 import i18n from '@cdo/locale';
-import CdoFieldDanceAi from './cdoFieldDanceAi';
+import CdoFieldDanceAi from './ai/cdoFieldDanceAi';
 
 // This color palette is limited to colors which have different hues, therefore
 // it should not contain different shades of the same color such as
